// src/Types/chapter1.ts
export interface chapter1 {
  language: string
  summary: string[]
  summary1: string[]
  summary2: string[]
  summary3: string[]
  summary4: string[]
  summary5: string[]
  summary6: string[]
  summary7: string[]
  summary8: string[]
  summary9: string[]
  summary10: string[]
  summary11:string[]
  summary12:string[]
}
export interface SummaryListProps {
  items: string[]
  className?:string

}